+++ CHANGELOG.md
@@ -2,12 +2,22 @@
 
 The changelog for `SuperwallKit`. Also see the [releases](https://github.com/superwall/Superwall-iOS/releases) on GitHub.
 
+## 4.0.0-beta.3
+
+### Breaking Changes
+
+- Renames `PaywallView(event:)` to `PaywallView(placement:)`.
+
+### Fixes
+
+- Adds extra check to get StoreKit 2 transaction data on `transaction_complete`.
+
 ## 4.0.0-beta.2
 
 ### Fixes
 
 - Fixes an issue to do with audience filters.
-- Readds unavailable functions from v3 to make the upgrade path smoother.
+- Re-adds unavailable functions from v3 to make the upgrade path smoother.
 
 ## 4.0.0-beta.1